Draft for activity - includede activity controller, repository and service

+++ src/main/java/com/eztix/eventservice/repository/ActivityRepository.java
@@ -3,8 +3,6 @@ package com.eztix.eventservice.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.eztix.eventservice.model.Activity;
 
-public interface ActivityRepository extends CrudRepository<Activity, Integer>{
-
+public interface ActivityRepository extends CrudRepository<Activity, Long>{
     //SQL -> SELECT * FROM ACTIVITY WHERE ID = {input_iD}
-
 }
